[chore] #4 shop_detail_view 하나의 activity로 합치기 (fragment 제거)
