+++ app/src/main/java/org/sopt/tabling/presentation/shopDetail/ShopDetailMenuListAdapter.kt
@@ -1,4 +1,4 @@
-package org.sopt.tabling.presentation.shopDetailMenuList
+package org.sopt.tabling.presentation.shopDetail
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
