+++ app/src/main/java/org/sopt/tabling/presentation/shopDetail/ShopDetailMenuViewHolder.kt
@@ -1,4 +1,4 @@
-package org.sopt.tabling.presentation.shopDetailMenuList
+package org.sopt.tabling.presentation.shopDetail
 
 import android.content.Context
 import androidx.recyclerview.widget.RecyclerView
